refactored cortex data and plots
