+++ dash_connectivity_viewer/common/neuron_data_base.py
@@ -1,117 +1,326 @@
+from logging import root
+from caveclient.chunkedgraph import root_id_int_list_check
 import pandas as pd
 import numpy as np
-import datetime
 from functools import lru_cache
-from typing import *
 from caveclient import CAVEclient
-import datetime
 
-import plotly.graph_objects as go
-from .client_patch import patch_client
+from dash_connectivity_viewer.common.lookup_utilities import (
+    get_nucleus_id_from_root_id,
+    get_root_id_from_nuc_id,
+)
+
 from .dataframe_utilities import *
 from .config import *
+from .link_utilities import voxel_resolution_from_info
+from multiprocessing import cpu_count
+
+
+def _soma_property_entry(soma_table, soma_id_column, soma_position_col, soma_filter):
+    return {
+        soma_table: {
+            "root_id": soma_id_column,
+            "include": [soma_position_col],
+            "aggregate": {
+                num_soma_col: {
+                    "group_by": soma_id_column,
+                    "column": NUCLEUS_ID_COLUMN,
+                    "agg": "count",
+                }
+            },
+            "suffix": "_soma",
+            "table_filter": soma_filter,
+            "data": None,
+        }
+    }
 
 
 class NeuronData(object):
     def __init__(
         self,
-        oid: int,
-        client: CAVEclient,
-        synapse_table: str = None,
-        cell_type_table: str = None,
-        soma_table: str = None,
-        timestamp: Union[datetime.datetime, None] = None,
-        axon_only: bool = False,
-        split_threshold: float = 0.7,
-        live_query: bool = True,
-    ) -> None:
-        self._oid = oid
-        self._client = patch_client(client)
+        object_id,
+        client,
+        property_tables={},
+        timestamp=None,
+        synapse_table=None,
+        soma_table=None,
+        n_threads=None,
+        synapse_position_point=syn_pt_position_col,
+        cell_position_point=cell_pt_position_col,
+        soma_id_column=NUCLEUS_ID_COLUMN,
+        id_type="root",
+        soma_table_query=soma_table_query,
+        num_soma_column=num_soma_col,
+    ):
 
-        self._live_query = live_query
+        if id_type == "root":
+            self._root_id = object_id
+            self._nucleus_id = None
+        elif id_type == "nucleus":
+            self._root_id = None
+            self._nucleus_id = object_id
 
-        if timestamp is None:
-            timestamp = datetime.datetime.now()
+        self._client = CAVEclient(
+            datastack_name=client.datastack_name,
+            server_address=client.server_address,
+            auth_token=client.auth.token,
+            pool_block=True,
+            pool_maxsize=2 * MAX_CHUNKS,
+        )
 
-        self._timestamp = timestamp
-        self.axon_only = axon_only
+        self._property_tables = property_tables
 
         if synapse_table is None:
-            synapse_table = client.info.get_datastack_info()["synapse_table"]
-        self.synapse_table = synapse_table
-        self.soma_table = soma_table
-        self.cell_type_table = cell_type_table
-        self.split_threshold = split_threshold
-        self._post_syn_df = None
+            synapse_table = client.info.get_datastack_info().get("synapse_table")
+        self._synapse_table = synapse_table
+
+        if soma_table is None:
+            soma_table = client.info.get_datastack_info().get("soma_table")
+
+        self._soma_table = soma_table
+        self._soma_id_column = soma_id_column
+        self._num_soma_column = num_soma_column
+
+        self._timestamp = timestamp
+
         self._pre_syn_df = None
+        self._post_syn_df = None
+
+        self._viewer_resolution = voxel_resolution_from_info(client.info.info_cache)
+
+        if n_threads is None:
+            n_threads = cpu_count()
+        self.n_threads = n_threads
+
+        self._synapse_position_point = synapse_position_point
+        self._cell_position_point = cell_position_point
+
+        self._partner_soma_table = None
+
+        if self._soma_table is not None:
+            self._property_tables.update(
+                _soma_property_entry(
+                    soma_table,
+                    self.cell_root_id_column,
+                    self.cell_position_column,
+                    soma_table_query,
+                )
+            )
 
     @property
-    def oid(self) -> int:
-        return self._oid
+    def synapse_position_column(self):
+        return bound_pt_position(self._synapse_position_point)
 
     @property
-    def client(self) -> CAVEclient:
+    def cell_position_column(self):
+        return bound_pt_position(self._cell_position_point)
+
+    @property
+    def cell_root_id_column(self):
+        return bound_pt_root_id(self._cell_position_point)
+
+    @property
+    def root_id(self):
+        if self._root_id is None:
+            new_root_id = get_root_id_from_nuc_id(
+                self._nucleus_id,
+                self.client,
+                self.soma_table,
+                self.cell_root_id_column,
+                self._soma_id_column,
+                self.timestamp,
+            )
+            if new_root_id is None:
+                raise Exception("Nucleus ID not found in soma table")
+            else:
+                self._root_id = new_root_id
+        return self._root_id
+
+    @property
+    def nucleus_id(self):
+        if self._nucleus_id is None:
+            self._nucleus_id = get_nucleus_id_from_root_id(
+                self._root_id,
+                self.client,
+                self.soma_table,
+                self.cell_root_id_column,
+                self._soma_id_column,
+                self.timestamp,
+            )
+        return self._nucleus_id
+
+    @property
+    def client(self):
         return self._client
 
     @property
-    def timestamp(self) -> datetime.datetime:
+    def live_query(self):
+        return self._timestamp is not None
+
+    @property
+    def timestamp(self):
         return self._timestamp
 
     @property
-    def live_query(self) -> bool:
-        return self._live_query
+    def synapse_table(self):
+        return self._synapse_table
 
-    def pre_syn_df(self) -> pd.DataFrame:
+    @property
+    def soma_table(self):
+        return self._soma_table
+
+    @property
+    def num_soma_column_full(self):
+        return f"{self._num_soma_column}{self.property_column_suffix(self.soma_table)}"
+
+    @property
+    def soma_position_column(self):
+        return (
+            f"{self.cell_position_column}{self.property_column_suffix(self.soma_table)}"
+        )
+
+    @property
+    def property_tables(self):
+        return [k for k in self._property_tables]
+
+    def pre_syn_df(self):
         if self._pre_syn_df is None:
             self._get_syn_df()
-        return self._pre_syn_df
+        return self._pre_syn_df.copy()
 
-    def post_syn_df(self) -> pd.DataFrame:
+    def post_syn_df(self):
         if self._post_syn_df is None:
             self._get_syn_df()
-        return self._post_syn_df
-
-    @lru_cache(maxsize=1)
-    def pre_targ_simple_df(self) -> pd.DataFrame:
-        syn_df_grp = self.pre_syn_df().groupby("post_pt_root_id")
-        targ_simple_df = self._make_simple_targ_df(syn_df_grp)
-        return targ_simple_df.rename(columns={"post_pt_root_id": "root_id"})
-
-    @lru_cache(maxsize=1)
-    def post_targ_simple_df(self) -> pd.DataFrame:
-        syn_df_grp = self.post_syn_df().groupby("pre_pt_root_id")
-        targ_simple_df = self._make_simple_targ_df(syn_df_grp)
-        return targ_simple_df.rename(columns={"pre_pt_root_id": "root_id"})
-
-    def _make_simple_targ_df(self, df_grp):
-        pts = df_grp["ctr_pt_position"].agg(list)
-        net_size = df_grp["size"].agg(sum)
-        mean_size = df_grp["size"].agg(np.mean)
-        num_syn = df_grp["ctr_pt_position"].agg(len)
-        return (
-            pd.DataFrame(
-                {
-                    "ctr_pt_position": pts,
-                    "num_syn": num_syn,
-                    "net_syn_size": net_size,
-                    "mean_syn_size": mean_size.astype(int),
-                }
-            )
-            .sort_values(by="num_syn", ascending=False)
-            .reset_index()
-        )
+        return self._post_syn_df.copy()
 
     def _get_syn_df(self):
         self._pre_syn_df, self._post_syn_df = synapse_data(
-            self.synapse_table,
-            self.oid,
+            synapse_table=self.synapse_table,
+            root_id=self.root_id,
+            client=self.client,
+            timestamp=self.timestamp,
+            live_query=self.live_query,
+            n_threads=self.n_threads,
+            synapse_position_column=self.synapse_position_column,
+        )
+        self._populate_root_ids()
+        self._populate_property_tables()
+
+    def _populate_root_ids(self):
+        self._partner_root_ids = np.unique(
+            np.concatenate(
+                (
+                    self._pre_syn_df["post_pt_root_id"].values,
+                    self._post_syn_df["pre_pt_root_id"].values,
+                )
+            )
+        )
+
+    def partners_out(self, properties=True):
+        return self._targ_table("pre", properties)
+
+    def partners_in(self, properties=True):
+        return self._targ_table("post", properties)
+
+    def _targ_table(self, side, properties):
+        if side == "pre":
+            prefix = "post"
+            syn_df_grp = self.pre_syn_df().groupby(f"{prefix}_pt_root_id")
+        elif side == "post":
+            prefix = "pre"
+            syn_df_grp = self.post_syn_df().groupby(f"{prefix}_pt_root_id")
+        targ_df = self._make_simple_targ_df(syn_df_grp).rename(
+            columns={f"{prefix}_pt_root_id": "root_id"}
+        )
+        if properties:
+            targ_df = self._merge_property_tables(targ_df, root_id_col)
+        return targ_df
+
+    def _make_simple_targ_df(self, df_grp):
+        pts = df_grp[self.synapse_position_column].agg(list)
+        net_size = df_grp["size"].agg(sum)
+        mean_size = df_grp["size"].agg(np.mean)
+        num_syn = df_grp[self.synapse_position_column].agg(len)
+        return (
+            pd.DataFrame(
+                {
+                    self.synapse_position_column: pts,
+                    num_syn_col: num_syn,
+                    net_size_col: net_size,
+                    mean_size_col: mean_size.astype(int),
+                }
+            )
+            .sort_values(by=num_syn_col, ascending=False)
+            .reset_index()
+        )
+
+    def _populate_property_tables(self):
+        dfs = property_table_data(
+            self._partner_root_ids,
+            self._property_tables,
+            self.client,
+            self.timestamp,
+            self.n_threads,
+        )
+        for k, df in dfs.items():
+            self._property_tables[k]["data"] = df
+
+    def property_data(self, table_name):
+        if self._property_tables.get(table_name).get("data") is None:
+            self._populate_property_tables()
+        return self._property_tables.get(table_name).get("data")
+
+    def property_root_id_column(self, table_name):
+        return self._property_tables.get(table_name).get("root_id")
+
+    def property_columns(self, table_name):
+        return [self.property_root_id_column(table_name)] + self._property_tables.get(
+            table_name
+        ).get("include")
+
+    def property_column_suffix(self, table_name):
+        return self._property_tables.get(table_name).get("suffix", "")
+
+    def _merge_property_tables(self, df, merge_column):
+        for tn in self.property_tables:
+            df = df.merge(
+                self.property_data(tn),
+                left_on=merge_column,
+                right_index=True,
+                how="left",
+                suffixes=("", self.property_column_suffix(tn)),
+            )
+            df.rename(
+                columns={
+                    c: f"{c}{self.property_column_suffix(tn)}"
+                    for c in self.property_columns(tn)
+                    if f"{c}{self.property_column_suffix(tn)}" not in df.columns
+                },
+                inplace=True,
+            )
+
+        if self.soma_table is not None:
+            df[self.num_soma_column_full] = (
+                df[self.num_soma_column_full].fillna(0).astype(int)
+            )
+
+        return df
+
+    def _get_own_soma_loc(self):
+        own_soma_df = get_specific_soma(
+            self.soma_table,
+            self.root_id,
             self.client,
             self.timestamp,
             live_query=self.live_query,
         )
+        if len(own_soma_df) != 1:
+            own_soma_loc = np.nan
+        else:
+            own_soma_loc = own_soma_df[self.cell_position_column].values[0]
+        return own_soma_loc
 
-    @lru_cache(maxsize=1)
-    def syn_df(self) -> pd.DataFrame:
+    def syn_all_df(self):
         pre_df = self.pre_syn_df()
         pre_df["direction"] = "pre"
         post_df = self.post_syn_df()
@@ -120,246 +329,11 @@ class NeuronData(object):
         syn_df["x"] = 0
         return syn_df
 
-    def _violin_plot(self, direction, name, side, color, xaxis, yaxis):
-        return go.Violin(
-            x=self.syn_df().query("direction == @direction")["x"],
-            y=self.syn_df().query("direction == @direction")[syn_depth_col],
-            side=side,
-            scalegroup="syn",
-            name=name,
-            points=False,
-            line_color=f"rgb{color}",
-            fillcolor=f"rgb{color}",
-            xaxis=xaxis,
-            yaxis=yaxis,
-        )
-
-    def post_violin_plot(
-        self,
-        color: Iterable[float],
-        xaxis: Union[str, None] = None,
-        yaxis: Union[str, None] = None,
-    ) -> go.Violin:
-        return self._violin_plot(
-            "post", "Post", "negative", color, xaxis=xaxis, yaxis=yaxis
-        )
-
-    def pre_violin_plot(
-        self,
-        color: Iterable[float],
-        xaxis: Union[str, None] = None,
-        yaxis: Union[str, None] = None,
-    ) -> go.Violin:
-        return self._violin_plot(
-            "pre", "Post", "positive", color, xaxis=xaxis, yaxis=yaxis
-        )
-
-    def _get_own_soma_loc(self):
-        own_soma_df = get_specific_soma(
-            self.soma_table,
-            self.oid,
-            self.client,
-            self.timestamp,
-            live_query=self.live_query,
-        )
-        if len(own_soma_df) != 1:
-            own_soma_loc = np.nan
-        else:
-            own_soma_loc = own_soma_df["pt_position"].values[0]
-        return own_soma_loc
-
     @lru_cache(maxsize=1)
-    def soma_location(self) -> np.ndarray:
+    def soma_location(self):
         if self.soma_table is None:
             return None
         return np.array(self._get_own_soma_loc())
 
-    def soma_location_list(self, length: int) -> list:
+    def soma_location_list(self, length):
         return np.repeat(np.atleast_2d(self.soma_location()), length, axis=0).tolist()
-
-    def _get_ct_soma_df(self, target_ids):
-        targ_ct_soma_df = cell_typed_soma_df(
-            self.soma_table,
-            self.cell_type_table,
-            target_ids,
-            self.client,
-            self.timestamp,
-            live_query=self.live_query,
-        )
-        return targ_ct_soma_df
-
-    def _target_ids(self):
-        pre_oids = self.pre_syn_df()["post_pt_root_id"]
-        post_oids = self.post_syn_df()["pre_pt_root_id"]
-        return np.unique(np.concatenate([pre_oids, post_oids]))
-
-    @lru_cache(maxsize=5)
-    def targ_soma_df(self):
-        return self._get_ct_soma_df(self._target_ids())
-
-    def _compute_pre_targ_df(self):
-        pre_df = self.pre_syn_df()
-        targ_soma_df = self.targ_soma_df()
-
-        pre_targ_df = pre_df.merge(
-            targ_soma_df,
-            left_on="post_pt_root_id",
-            right_on="pt_root_id",
-            how="left",
-        ).drop(columns=["pt_root_id"])
-        pre_targ_df[num_soma_col].fillna(0, inplace=True)
-        pre_targ_df[num_soma_col] = pre_targ_df[num_soma_col].astype(int)
-
-        pre_targ_df[own_soma_col] = self.soma_location_list(len(pre_targ_df))
-        if len(pre_targ_df) == 0:
-            pre_targ_df[soma_dist_col] = []
-        else:
-            pre_targ_df[soma_dist_col] = pre_targ_df.apply(
-                lambda x: radial_distance(
-                    x, soma_position_col, own_soma_col, voxel_resolution
-                ),
-                axis=1,
-            )
-        return pre_targ_df
-
-    @lru_cache(maxsize=5)
-    def pre_targ_df(self) -> pd.DataFrame:
-        return self._compute_pre_targ_df().fillna(np.nan)
-
-    def _compute_post_targ_df(self):
-        post_df = self.post_syn_df()
-        targ_soma_df = self.targ_soma_df()
-
-        post_targ_df = post_df.merge(
-            targ_soma_df,
-            left_on="pre_pt_root_id",
-            right_on="pt_root_id",
-            how="left",
-        ).drop(columns=["pt_root_id"])
-        post_targ_df[num_soma_col].fillna(0, inplace=True)
-        post_targ_df[num_soma_col] = post_targ_df[num_soma_col].astype(int)
-        post_targ_df[own_soma_col] = self.soma_location_list(len(post_targ_df))
-        if len(post_targ_df) == 0:
-            post_targ_df[soma_dist_col] = []
-        else:
-            post_targ_df[soma_dist_col] = post_targ_df.apply(
-                lambda x: radial_distance(
-                    x, soma_position_col, own_soma_col, voxel_resolution
-                ),
-                axis=1,
-            )
-        return post_targ_df
-
-    @lru_cache(maxsize=5)
-    def post_targ_df(self) -> pd.DataFrame:
-        return self._compute_post_targ_df().fillna(np.nan)
-
-    def synapse_soma_scatterplot(
-        self,
-        valence_colors: dict,
-        xaxis: Union[str, None] = None,
-        yaxis: Union[str, None] = None,
-    ) -> go.Scattergl:
-        pre_targ_df = self.pre_targ_df().dropna(
-            subset=[syn_depth_col, soma_depth_col, valence_col]
-        )
-        if len(pre_targ_df) > 0:
-            color_vec = pre_targ_df[valence_col].astype(int)
-        else:
-            color_vec = 0
-        return go.Scattergl(
-            x=pre_targ_df[soma_depth_col],
-            y=pre_targ_df[syn_depth_col],
-            mode="markers",
-            marker=dict(
-                color=valence_colors[color_vec],
-                line_width=0,
-                size=5,
-                opacity=0.5,
-            ),
-            xaxis=xaxis,
-            yaxis=yaxis,
-        )
-
-    def bar_data(self) -> pd.Series:
-        pre_targ_df = self.pre_targ_df().dropna(subset=["cell_type"])
-        return pre_targ_df.groupby(ct_col).count()["size"]
-
-    def _bar_plot(self, name, indices, color):
-        return go.Bar(
-            name=name,
-            y=self.bar_data().loc[indices].index,
-            x=self.bar_data().loc[indices].values,
-            marker_color=f"rgb{color}",
-            orientation="h",
-        )
-
-    def excitatory_bar_plot(self, color: Iterable[float]) -> go.Bar:
-        return self._bar_plot(
-            "Exc. Targets", exc_types, tuple(np.floor(255 * color).astype(int))
-        )
-
-    def inhib_bar_plot(self, color: Iterable[float]) -> go.Bar:
-        return self._bar_plot(
-            "Inh. Targets", inhib_types, tuple(np.floor(255 * color).astype(int))
-        )
-
-    def _compute_tab_dat(self, direction):
-        if direction == "pre":
-            df = self.pre_targ_df()
-            merge_column = "post_pt_root_id"
-        elif direction == "post":
-            df = self.post_targ_df()
-            merge_column = "pre_pt_root_id"
-        if len(df) == 0:
-            df["ctr_pt_position"] = []
-            df[num_syn_col] = []
-            df[net_size_col] = []
-            df[mean_size_col] = []
-        else:
-            df[num_syn_col] = df.groupby(merge_column).transform("count")[
-                "ctr_pt_position"
-            ]
-            df[net_size_col] = (
-                df[[merge_column, "size"]]
-                .groupby(merge_column)
-                .transform("sum")["size"]
-            )
-            df[mean_size_col] = (
-                df[[merge_column, "size"]]
-                .groupby(merge_column)
-                .transform("mean")["size"]
-            ).astype(int)
-
-        df_unique = (
-            df.drop_duplicates(subset=merge_column)
-            .drop(columns=["size"])
-            .drop(columns="ctr_pt_position")
-        )
-
-        pt_df = df.groupby(merge_column)["ctr_pt_position"].agg(list)
-        df_unique = df_unique.merge(pt_df, left_on=merge_column, right_index=True)
-
-        tab_dat = df_unique.sort_values(by=num_syn_col, ascending=False)
-        tab_dat[merge_column] = tab_dat[merge_column].astype(
-            str
-        )  # Dash can't handle int64
-        return tab_dat
-
-    @lru_cache(maxsize=5)
-    def pre_tab_dat(self) -> pd.DataFrame:
-        return (
-            self._compute_tab_dat("pre")
-            .fillna(np.nan)
-            .drop(columns=["pre_pt_root_id"])
-            .rename(columns={"post_pt_root_id": "root_id"})
-        )
-
-    @lru_cache(maxsize=5)
-    def post_tab_dat(self) -> pd.DataFrame:
-        return (
-            self._compute_tab_dat("post")
-            .fillna(np.nan)
-            .drop(columns=["post_pt_root_id"])
-            .rename(columns={"pre_pt_root_id": "root_id"})
-        )
